Enum gluon_check::typecheck::TypeError

pub enum TypeError<I> {
    UndefinedVariable(I),
    NotAFunction(ArcType<I>),
    UndefinedType(I),
    UndefinedField(ArcType<I>, I),
    PatternError(ArcType<I>, usize),
    Unification(ArcType<I>, ArcType<I>, Vec<Error<I>>),
    KindError(Error<I>),
    Rename(RenameError),
    DuplicateTypeDefinition(I),
    DuplicateField(I),
    InvalidProjection(ArcType<I>),
    UndefinedRecord {
        fields: Vec<I>,
    },
    EmptyCase,
}

Type representing a single error when checking a type

Variants
UndefinedVariable(I)

Variable has not been defined before it was used

NotAFunction(ArcType<I>)

Attempt to call a type which is not a function

UndefinedType(I)

Type has not been defined before it was used

UndefinedField(ArcType<I>, I)

Type were expected to have a certain field

PatternError(ArcType<I>, usize)

Constructor type was found in a pattern but did not have the expected number of arguments

Unification(ArcType<I>, ArcType<I>, Vec<Error<I>>)

Errors found when trying to unify two types

KindError(Error<I>)

Error were found when trying to unify the kinds of two types

Rename(RenameError)

Errors found during renaming (overload resolution)

DuplicateTypeDefinition(I)

Multiple types were declared with the same name in the same expression

DuplicateField(I)

A field was defined more than once in a record constructor or pattern match

InvalidProjection(ArcType<I>)

Type is not a type which has any fields

UndefinedRecord

Expected to find a record with the following fields

Fields

fields: Vec<I>

EmptyCase

Found a case expression without any alternatives