Trait gluon_check::unify::Unifier

pub trait Unifier<S, Type>: Sized where Type: Unifiable<S> {
    fn report_error(unifier: &mut UnifierState<S, Self>, error: Error<Type, Type::Error>);
    fn try_match(unifier: &mut UnifierState<S, Self>, l: &Type, r: &Type) -> Option<Type>;
}

A Unifier is a type which implements a unifying strategy between two values.

Required Methods

fn report_error(unifier: &mut UnifierState<S, Self>, error: Error<Type, Type::Error>)

Reports an error to the unifier for cases when returning the error is not possible.

fn try_match(unifier: &mut UnifierState<S, Self>, l: &Type, r: &Type) -> Option<Type>

Attempt to unify l and r using the strategy of Self.